fix: adjust component
