+++ package.json
@@ -42,7 +42,7 @@
     "react-dom": "^18.3.1",
     "react-router-dom": "^6.24.1",
     "styled-components": "^6.1.11",
-    "yup": "^1.2.0"
+    "yup": "^1.4.0"
   },
   "devDependencies": {
     "@chromatic-com/storybook": "^1.6.1",
